add network middleware test and move files
