+++ src/store/middlewares/network.js
@@ -1,14 +1,12 @@
-import { networkConfigSet } from '../../actions/network';
+import { networkConfigSet } from '../../actions';
 import actionsType from '../../constants/actions';
 
 const network = ({ dispatch }) => next => async (action) => {
   switch (action.type) {
     case actionsType.networkSelected: {
-      console.log('networkMiddleware', action.data);
       dispatch(await networkConfigSet(action.data));
       break;
     }
-
     default:
       break;
   }
